Update reference spaces to new model
