+++ components/script/dom/webidls/XRReferenceSpace.webidl
@@ -4,6 +4,14 @@
 
 // https://immersive-web.github.io/webxr/#xrreferencespace-interface
 
+enum XRReferenceSpaceType {
+  "viewer",
+  "local",
+  "local-floor",
+  "bounded-floor",
+  "unbounded"
+};
+
 [SecureContext, Exposed=Window, Pref="dom.webxr.enabled"]
 interface XRReferenceSpace : XRSpace {
   attribute XRRigidTransform originOffset;
